+++ src/adapter/cloudflare-workers/index.ts
@@ -1,2 +1,2 @@
 // @denoify-ignore
-export { serveStatic } from './server-static-module'
+export { serveStatic } from './serve-static-module'
