+++ src/adapter/cloudflare-workers/serve-static-module.ts
